QOL: Code Refactored into seperate folders and files
