+++ Lectra/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "ChatGPT Image Mar 29, 2025, 07_22_47 PM.png",
+      "filename" : "SelahNotes-Logo.png",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
